+++ src/Framework/Domain/FilterInterceptor.php
@@ -47,7 +47,7 @@ class FilterInterceptor implements CoreInterceptorInterface
     /**
      * @inheritDoc
      */
-    public function process(string $controller, string $action, array $parameters, CoreInterface $core)
+    public function process(string $controller, string $action, array $parameters, CoreInterface $core): mixed
     {
         foreach ($this->getDeclaredFilters($controller, $action) as $parameter => $filterClass) {
             if (isset($parameters[$parameter])) {